rs232_agent/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyserial==3.5
Werkzeug==3.1.3
zipp==3.22.0
tkinter
websockets==11.0.3
